### 2018-07-28 ed27480
Added missing comma for BibTeX.
