+++ JOSS-paper/paper.bib
@@ -12,7 +12,7 @@
     journal = {Proc Natl Acad Sci USA},
     volume = {108},
     number = {17},
-    pages = {7265-7270}
+    pages = {7265-7270},
     doi = {10.1073/pnas.1102826108}
 }
 
