### 2018-07-28 #858 @rrrlw
Thank you, @arfon, for the quick response and all the efforts put behind making JOSS successful. I suggested @leeper as the corresponding bio mentioned extensive experience with R packages. @leeper, I hope you find this an interesting submission. I have two notes that should make the reviewers'/editors' tasks easier:

1. TDAstats v0.2.0 (development version on GitHub repo) is being submitted to JOSS, even though only v0.1.0 is currently available on CRAN. Assuming all goes well, I was hoping to submit the final version of the package to CRAN and, given CRAN's submission policy (update every 1-2 months), would prefer to avoid overburdening CRAN volunteers by also submitting v0.2.0 (version would be incremented to 1.0.0 post-review). Thus, instead of using `install.packages("TDAstats")`, reviewers would need to use `devtools::install_github("rrrlw/TDAstats")` to evaluate the most recent version of the package.
2. The markdown file at https://github.com/rrrlw/TDAstats/blob/master/JOSS-paper/reviewer-checklist.md contains the JOSS reviewer checklist and how I think this submission meets each point. I hope it saves the reviewers' time.

Thank you, again, for your time. @whedon appears to have found a formatting error in the bibliography file. I will correct it ASAP, and recompile the pdf.

Best,
Raoul
