### 2018-07-28 #858 @rrrlw
The error has been fixed, and the article proof generated by @whedon appears as intended. From the list of people who have agreed to review for JOSS, @corybrunson would likely be the best fit to review this submission given the familiarity with R and expertise in topological data analysis. Please let me know if there's anything I can do to make the pre-review/review process easier. Thank you!
