### 2018-07-28 1fabc1b
Minor fix for JOSS submission's reviewer checklist.
