### 2018-07-28 #858 @corybrunson
I would be delighted to review this package! I can finish within one week. (I wouldn't call my experience expertise, but i'm familiar with the concepts implemented.)
