### 2018-07-29 #860 @whedon
[REVIEW]: TDAstats: R pipeline for computing persistent homology in topological data analysis
**Submitting author:** @rrrlw (<a href="http://orcid.org/0000-0003-0503-9580">Raoul R. Wadhwa</a>)
**Repository:** <a href="https://github.com/rrrlw/TDAstats" target ="_blank">https://github.com/rrrlw/TDAstats</a>
**Version:** v0.2.0
