### 2018-07-29 d467905
Added JOSS status badge.
