### 2018-07-31 fbc97ce
Added back C++ method for reading lower distance matrix (but from R numeric matrix).
