+++ src/ripser_short.cpp
@@ -549,9 +549,30 @@ compressed_lower_distance_matrix getPointCloud(NumericMatrix inputMat) {
   return compressed_lower_distance_matrix(std::move(distances));
 }
 
-// More lightweight version of Ripser by Ulrich Bauer
+compressed_lower_distance_matrix getLowerDistMatrix(NumericMatrix inputMat) {
+	std::vector<value_t_ripser> distances;
+	value_t_ripser value;
+
+	int numRows = inputMat.nrow(),
+		numCols = inputMat.ncol();
+
+	for (int i = 0; i < numRows; i++)
+	{
+		for (int j = 0; j < i; j++)
+		{
+			value = inputMat(i, j);
+			distances.push_back(value);
+		}
+	}
+
+	return compressed_lower_distance_matrix(std::move(distances));
+}
+
+// Altered version of Ripser by Ulrich Bauer
+// format = 0 --> point cloud
+// format = 1 --> lower distance matrix
 // [[Rcpp::export]]
-NumericVector ripser_cpp(NumericMatrix input_points, int dim, float thresh) {
+NumericVector ripser_cpp(NumericMatrix input_points, int dim, float thresh, int format) {
   NumericVector ansx(1);
   ansx[0] = 1;
 
