### 2018-07-31 8ceab72
Added lower distance matrix format to main method.
