### 2018-07-31 59191a7
Modified R code for proper C++ linking via Rcpp (added format parameter and management).
