### 2018-07-31 0305776
Fixed issue with reading lower distance matrix in C++ code, and updated documentation.
