+++ man/calculate_homology.Rd
@@ -4,14 +4,16 @@
 \alias{calculate_homology}
 \title{Calculate Persistent Homology of a Point Cloud}
 \usage{
-calculate_homology(mat, dim = 1, threshold = -1)
+calculate_homology(mat, dim = 1, threshold = -1, format = "cloud")
 }
 \arguments{
-\item{mat}{numeric matrix containing point cloud}
+\item{mat}{numeric matrix containing point cloud or lower distance matrix}
 
 \item{dim}{maximum dimension of features to calculate}
 
 \item{threshold}{maximum diameter for computation of Vietoris-Rips complexes}
+
+\item{format}{format of `mat`, either "cloud" for point cloud or "ldm" for lower distance matrix}
 }
 \value{
 3-column matrix, with each row representing a TDA feature
@@ -27,6 +29,8 @@ The `mat` parameter should be a numeric matrix with each row corresponding
 to a single point, and each column corresponding to a single dimension. Thus,
 if `mat` has 50 rows and 5 columns, it represents a point cloud with 50 points
 in 5 dimensions. The `dim` parameter should be a positive integer.
+Alternatively, the `mat` parameter could be a lower distance matrix (upper
+triangular half is ignored); note: `format` should be specified as "ldm".
 }
 \examples{
 
