### 2018-08-01 3d5e4c5
Updated documentation: ldm --> distmat (regular distance matrix, less work for user).
