### 2018-08-01 f45717a
Added format parameter to permutation_test.
