### 2018-08-01 952ac5c
Added update parameter to permutation_test.
