### 2018-08-01 5176df1
Updated documentation for permutation_test (update parameter).
