+++ man/permutation_test.Rd
@@ -5,7 +5,7 @@
 \title{Statistical Inference for Topological Data Analysis}
 \usage{
 permutation_test(data1, data2, iterations, exponent = 1, dim = 1,
-  format = "cloud", threshold = -1)
+  format = "cloud", threshold = -1, update = 0)
 }
 \arguments{
 \item{data1}{first dataset}
@@ -21,6 +21,8 @@ permutation_test(data1, data2, iterations, exponent = 1, dim = 1,
 \item{format}{format of data, either "cloud" for point cloud or "distmat" for distance matrix}
 
 \item{threshold}{maximum diameter for computation of Vietoris-Rips complexes}
+
+\item{update}{if greater than zero, will print a message every `update` iterations}
 }
 \value{
 list containing results of permutation test
