### 2018-08-01 28dc625
Added MANY error messages in permutation_test to debug long runtimes with distance matrix format.
