### 2018-08-02 cba0019
Ensure that point cloud format is there for permutation_test.
