### 2018-08-02 4244307
Used ... instead of explicit parameters for calculate_homology passed first to permutation_test.
