### 2018-08-02 801d0bf
Added documentation for ellipsis in permutation_test.
