### 2018-08-02 4ccee41
Added function that measures distance between homology matrices (previously known as Wasserstein, now users can use it directly).
