### 2018-08-04 #860 @corybrunson
@rrrlw yes indeed. I'm sorry for the delay, as i had some deadlines to meet when i returned to work this week, but i intend to conduct my review today. Hopefully i won't have to leave anything for tomorrow. Thanks for checking!
