### 2018-08-04 #860 @rrrlw
All good, thank you again for reviewing!

A couple of quick notes: (1) the file at https://github.com/rrrlw/TDAstats/blob/master/JOSS-paper/reviewer-checklist.md might save you some time (adapted reviewer checklist from JOSS website); (2) `install.packages("TDAstats")` won't download the latest version of TDAstats so `devtools::install_github("rrrlw/TDAstats")` is required. Let me know if I can help in any way.
