### 2018-08-04 #860 @corybrunson
@rrrlw since you're the only author to have committed changes to the repo, could you comment on the contributions of your coauthors? I don't question the significance of their contributions; i just would like to have a point of reference for the last item under "General checks" above.
