### 2018-08-04 rrrlw/TDAstats#1 @corybrunson
real-world illustration
The toy examples provided in with the package are helpful and wholly appropriate. The package would also benefit, i think, from an illustration using real-world data (as suggested but not required by the JOSS review checklist). Is there a dataset you've used for this purpose that could be included in the package and demonstrated either in the functional documentation or in a separate vignette? This isn't a sticking point for my review, but i do think the package would achieve its aims more effectively with such a case study.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/860)._
