### 2018-08-04 rrrlw/TDAstats#2 @corybrunson
fixed coordinates for persistence plot?
Since persistence plots implicitly rely on a 1:1 aspect ratio for visual interpretability, would it be appropriate to include `+ ggplot2::coord_fixed(ratio = 1)` to the `'ggplot'` object returned by `plot_persist()`? (Certainly lower- and higher-persistence features are discriminable regardless of the aspect ratio, but for professional publications this has, to my knowledge, been the rule. And, like the other plot specs, this could be overridden by the user.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/860)._
