### 2018-08-04 rrrlw/TDAstats#3 @corybrunson
diagonal missing from modified persistence plot
The following code chunk doesn't produce an error, but fails (for me) to produce the diagonal line in the persistence plot (i had trouble using `reprex::reprex()`):

```r
library(TDAstats)
data(circle2d)
circ.phom <- calculate_homology(circle2d, dim = 1)
plot_persist(circ.phom) + ggplot2::xlim(c(0, .25))
```

Is the problem on my end? It prevents me from fully performing the visual comparison suggested in the "inference" vignette. I will try to reproduce it on a different machine tomorrow.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/860)._
