### 2018-08-04 #860 @corybrunson
@rrrlw i've raised all of the issues that have come up for me, so, as soon as they're resolved (some are just questions), my review will be done. The package is a great help—i only recently was pointed to Ripser and made do with some ad hoc functions to interface with it through R, which i'm glad to no longer have to do. : ) Thank you for the contribution!
