### 2018-08-04 fc3da1f
Fixed issue where diagonal wouldn't appear on persistence diagram with smaller xlim. This fixes issue #3.
