+++ R/visuals.R
@@ -53,7 +53,8 @@ plot_persist <- function(feature.matrix) {
                             y = c(0, axes.max))
   ggplot2::ggplot(data = feature.df) +
     ggplot2::xlim(axes.min, axes.max) + ggplot2::ylim(axes.min, axes.max) +                           # axis limits
-    ggplot2::geom_path(data = df.geompath, ggplot2::aes_string(x = "x", y = "y")) +          # reference segment
+    ggplot2::geom_abline(slope = 1, intercept = 0) +  # reference line
+    #ggplot2::geom_path(data = df.geompath, ggplot2::aes_string(x = "x", y = "y")) +          # reference segment
     ggplot2::xlab("Feature Birth") + ggplot2::ylab("Feature Death") +                                 # axis titles
     ggplot2::theme(axis.line = ggplot2::element_line(colour = "black"),                               # add axis lines
              panel.grid.major = ggplot2::element_blank(), panel.grid.minor = ggplot2::element_blank(),         # remove gridlines
