### 2018-08-04 1543117
Added 1:1 coordinate ratio for persistence diagrams. This fixes issue #2.
