### 2018-08-04 rrrlw/TDAstats#2 @rrrlw
Yes, you are correct, thank you for the suggestion! The fixed coordinate ratio has been added in commit 1543117.
