### 2018-08-04 rrrlw/TDAstats#1 @rrrlw
Yes, this is a great suggestion. Unfortunately, this package was created to cement the methodology of future research, rather than to solve a problem noticed in a prior project. As such, we do not yet have real-world data to which we have applied the functions in this package. I will leave the issue open, and when we do apply this method to real-world data with successful results, I will add the dataset and corresponding vignette/case study to the package.
