### 2018-08-04 d053abe
Added code showing how to install package with vignettes (thanks to @corybrunson for suggestion).
