### 2018-08-05 #860 @rrrlw
I added a separate line in README.md with `build_vignettes = TRUE`, and have replied to the 3 issues brought up in the TDAstats repo.

In terms of author contributions:
* Drew F.K. Williamson: software design, software testing, writing paper
* Andrew Dhawan: software design, software testing, writing paper
* Jacob G. Scott: software design, software testing, writing paper

Here, software design refers to selection of user-available functions, designing and selecting algorithms for implementation, providing expert opinion on the theoretical aspects of topological data analysis, and the aesthetic aspects of the visualizations. Software testing refers to manual testing of software features as they were implemented, for accuracy and efficiency (TDAstats and its previous iteration, ripserr). Writing the paper refers to preparation of the markdown file for submission to JOSS.

Thank you for the feedback! Did I miss responding to anything?
