### 2018-08-05 #860 @rrrlw
Question for @leeper: is it appropriate to add a short "Acknowledgments" section?
