### 2018-08-05 #860 @rrrlw
P.S. - I am really glad that you like TDAstats! I hope it saves you (and others) time/effort with TDA projects.
