### 2018-08-05 #860 @corybrunson
@rrrlw i'm satisfied! In particular, the most recent commit has resolved the diagonal line issue for me. Thanks for your thorough responses.

@leeper i have concluded my review and recommend acceptance of the current development version.
