### 2018-08-06 #860 @leeper
Thanks, @corybrunson! Much appreciated.
