### 2018-08-06 #860 @rrrlw
Will do, @leeper. The PR has been merged, I would just like to proofread paper.md once more. I will update you in this issue once that and steps 2-4 have been completed. Thank you!
