### 2018-08-07 0be77c6
Corrected acknowledgment and minor error with figure ref.
