+++ JOSS-paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 # Summary
 
-High-dimensional datasets are becoming more common in a variety of scientific fields. Well-known examples include next-generation sequencing in biology, patient health status in medicine, and computer vision in deep learning. Dimension reduction, using methods like principal component analysis (PCA), is a common preprocessing step for such datasets. However, while dimension reduction can save computing and human resources, it comes with the cost of significant information loss. Topological data analysis (TDA) aims to analyze the "shape" of high-dimensional datasets, without dimension reduction, by extracting features that are robust to small perturbations in data. Persistent features of a dataset can be used to describe it, and to compare it to other datasets. Visualization of persistent features can be done using topological barcodes or persistence diagrams (see figure). Application of TDA methods has granted greater insight into high-dimensional data [@tda-immune]; one prominent example of this is its use to characterize a clinically relevant subgroup of breast cancer patients [@tda-cancer].
+High-dimensional datasets are becoming more common in a variety of scientific fields. Well-known examples include next-generation sequencing in biology, patient health status in medicine, and computer vision in deep learning. Dimension reduction, using methods like principal component analysis (PCA), is a common preprocessing step for such datasets. However, while dimension reduction can save computing and human resources, it comes with the cost of significant information loss. Topological data analysis (TDA) aims to analyze the "shape" of high-dimensional datasets, without dimension reduction, by extracting features that are robust to small perturbations in data. Persistent features of a dataset can be used to describe it, and to compare it to other datasets. Visualization of persistent features can be done using topological barcodes or persistence diagrams (Figure 1). Application of TDA methods has granted greater insight into high-dimensional data [@tda-immune]; one prominent example of this is its use to characterize a clinically relevant subgroup of breast cancer patients [@tda-cancer].
 
 The ``TDAstats`` R package is a comprehensive pipeline for conducting TDA. Once data is loaded into R, ``TDAstats`` can calculate, visualize, and conduct nonparametric statistical inference on persistent homology. The Ripser C++ library [@Ripser], benchmarked at approximately 40 times faster than comparable software, is wrapped using Rcpp [@rcpp-paper] for efficient computation of persistent homology. ``TDAstats`` generates topological barcodes and persistence diagrams using the ubiquitous ggplot2 library [@ggplot2-book], allowing use of ggplot2 functions to manipulate plots. This reduces the number of manual steps required to prepare publication-quality figures, thus enabling reproducible research [@reproduce-res]. ``TDAstats`` also implements nonparametric hypothesis testing of persistent homology using a permutation test, first described by @tda-testing. To our knowledge, ``TDAstats`` is the first library to implement this feature.
 
@@ -44,6 +44,6 @@ The primary barrier to using TDA is not mathematical comprehension. Although the
 
 # Acknowledgments
 
-We thank Francesca Buffa, Ph.D. for introducing us to Ripser; Chad Topaz, Ph.D. for assistance as we started studying topological data analysis; J. Cory Brunson, Ph.D. and Jessica Scarborough, M.S. for insightful feedback.
+We thank Bernadette J. Stolz for introducing us to Ripser; Chad Topaz, Ph.D. for assistance as we started studying topological data analysis; J. Cory Brunson, Ph.D. and Jessica Scarborough, M.S. for insightful feedback.
 
 # References
