### 2018-08-07 44bea54
Added details on Nicolau et al study.
