### 2018-08-07 97521c4
Correctly formatted text referring to R code/packages.
