### 2018-08-07 d7268ed
Corrected acknowledgment; added details re: features.
