### 2018-08-08 68af578
Added code coverage to project.
