+++ .travis.yml
@@ -2,3 +2,5 @@
 language: r
 warnings_are_errors: true
 sudo: required
+after_success:
+- Rscript -e 'covr::codecov()'
