### 2018-08-08 #860 @rrrlw
@leeper, the current version of the PDF looks good, and TDAstats v0.3.2 has been archived on Zenodo with DOI: 10.5281/zenodo.1341719. Please let me know if there is anything else I need to do. Thank you!
